requirements.txt
requests>=2.31.0
PyGithub>=2.1.1
pytest>=7.4.0
